day 10 solution
This one took a lot of re-reading to understand, but it eventually got
there. For part 1, I used a simple loop structure and made sure I was
implementing the problem correctly.
Part 2, that same x logic applied, and all it needed was a bit more
index logic, which wasn't too bad.