Day 11 and 12 answers
(well, 12 kind of, but I got em)
11 took me a while to realize the trick (over a day, and I needed a GOOD
hint about divisors lol)
but 12, the answer only kind of works. My A* implementation is greedier
than I want, and I think I have an idea why, but I don't wanna be
bothered with fixing it right now. So, I visually inpected my result to
correct it and was able to do so relatively easily.
Part 2, I was also able to do visually.